It has come to my attention that I had accidentally pushed the wrong tree here that was based on an older version. This is the correct 5.08 tree. Updated the resources submodule as well. Including the new installer as well with the skeleton directory tree that is used in the VS solution to stage precompiled libs/plugins etc to the output dir (the final package directory tree basically) tha the installer uses as source. This may not be of use to a lot of people (as well as the new inno installer) but trying to avoid pushing these to the repo is how the wrong tree was pushed to begin with. It easier for me not to have to deal with separating things in order to keep this updated. I will merge the commits from stock after 5.08, once time allows again.

+++ Mapping.cpp
@@ -205,7 +205,7 @@ void CMUSHclientDoc::AddToMap (CString str)
       DrawMappingStatusLine ();
       } // end of finding the direction
 
-    str = str.Mid (iNewline + strlen (ENDLINE));
+    str = str.Mid (iNewline + (int)strlen (ENDLINE));
     }   // end of breaking the text into lines
 
   }   // end of CMUSHclientDoc::AddToMap
